+++ examples/pg_devsetup/bootstrap.sql
@@ -1,0 +1,5 @@
+create extension is_jsonb_valid;
+create extension pldbgapi;
+
+-- list the extensions
+\dx