+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Changes for docker-postgres
 
+## unreleased
+
+### Feature
+
+- added debugger extension
+
 ## 2020-04-29 / 0.1.0
 
 - Initial Release using Postgresql 12.2